% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepTree.R
\name{sweepTree}
\alias{sweepTree}
\title{Sweep Tree Function}
\usage{
sweepTree(V, M, df = NCOL(V))
}
\description{
\code{sweepTree} performs multiple sequences of sweep operations. To increase
computational stability no reverse sweeps are performed.  Instead this function
uses a binary tree and caching to reduce computational burden.
}

