% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.R
\name{isr}
\alias{isr}
\title{Iterative Sequential Regression}
\usage{
isr(X, M, mcmcIter = 100, sampleRate = 20, intercept = T)
}
\arguments{
\item{X}{A matrix of points to be imputed or used for covariates for the
by isr.  \code{NA} values are considered missing.  Distinct column names 
are required for each variable.}

\item{M}{A matrix specifying the relationships between each of the variables.
Each column name of M must match to the column names of X.  Each row of 
M identifies a variable that will be imputed for, all other variables are
treated as covariates for all variables.  Only the lower diagonal of this
matrix is used to identify relationships, relationships between the 
variable and itself are also ignored.  If missing, dependence is assumed
between all variables.}

\item{mcmcIter}{A scalar indicating the number of iterations to run.}

\item{sampleRate}{A scalar that provides thinning for the MCMC routine.}

\item{intercept}{A logical value identifying if the imputation model should 
have an intercept.}
}
\description{
\code{isr} performs imputation of missing values based on an optionally
specified model.  Missingness is assumed to be missing at random (MAR).
}
\examples{
set.seed(100)
mcmciter <- 100 

# these are dependent variables that we will record model parameters for
varList <-  c("Var1", "Var2", "Var3")

# simulation parameters
set.seed(100)
n <- 30
p <- 5

# generate a covar matrix
covarMatrix <- rWishart(1,p+1,diag(p))[,,1]

# simulation of variables under the variable relationships
U <- chol(covarMatrix)

X <- matrix(rnorm(n*p), nrow=n) \%*\% U
colnames(X) <- covarList

# specify relationships
fitMatrix <- matrix( c( 
#             Var1  Var2  Var3
# 1. Var1
                F,    F,    F,
# 2. Var2
                T,    F,    F,
# 3. Var3
                T,    T,    F 
),nrow=3,byrow=T)

colnames(fitMatrix) <- covarList 
rownames(fitMatrix) <- varList 

E<-  isr(X,fitMatrix,mcmcIter=mcmciter)

}
\references{
Robbins, M. W., & White, T. K. (2011). Farm commodity payments and imputation in the Agricultural Resource Management Survey. American journal of agricultural economics, DOI: 10.1093/ajae/aaq166.
}

